feat: Separate mappers and moved service
