+++ personalization-sdk/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
         </service>
-        <service android:name=".features.notification.data.service.NotificationService" />
+        <service android:name=".features.notification.service.NotificationService" />
 
     </application>
 </manifest>
