+++ personalization-sdk/src/main/kotlin/com/personalization/features/notification/event/mapper/EventType.mapper.kt
@@ -1,4 +1,4 @@
-package com.personalization.features.notification.data.mapper
+package com.personalization.features.notification.event.mapper
 
 import com.personalization.errors.JsonResponseErrorHandler
 import com.personalization.features.notification.domain.model.NotificationConstants.NOTIFICATION_PAYLOAD
